requirements.txt
praw==7.7.1
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==1.3.1
tweepy==4.14.0